Add minimum OS version check
Add support for checking both OS name and version.
Exit if an unsupported OS is detected, or if the version doesn't
meet the minimum version requirement specified.

Unfortunately we can't use the nice associative array syntax from
bash 4+ (macOS is stuck on bash 3 for licensing reasons…), so we
use a slightly less elegant workaround with variable name composition.