Add tmux key binding for synchronize-panes command
I can never remember the command `setw synchronize-panes` so rarely
use the feature, although I have had plenty of occasion to use it. After
searching for it the first few times I stopped using it as it was just
slowing me down. I meant to add a key binding but forgot, here it is at
last… Also updates the window borders and labels in status bar to make
it clear which windows are in sync mode.